added files from empty example.
Note: project must be placed in (OF-root)/apps/(app group name)/macFlock for the project to build. This restriction appears to be typical for OF projects built in XCode (header and library paths appear to be relative, the alternative could be a parent project managing everything OF-related, or defining environment variables set by a script during install).