Improvements.
This example of flocking behavior is dependent on the size of the volume the "birds" are contained in. I had previously scaled up the box size to better fit the OF view settings, I hadn't scaled up the other relevant parameters proportionally as well and this led to undesired behavior.

Problems still exist in the rotation function, but it is narrowed down to the lines for roll. Yaw and pitch now work correctly.