Moved flock #define's into the classes as member variables.
As part of the original sample code, #define's were used for adjustable parameters to encourage experimentation. This had the downside in that all flocks then had to use the same parameters.

There still exist several #define's in the Boid's class definition file. Some of these can potentially change to member variables as well.